Clean up selector utils and restructure files
